# openapi.yaml
openapi: 3.1.1
info:
  title: Dream Diary API + MCP
  version: 1.2.0
  description: REST API and MCP endpoint for creating and managing dream entries with OAuth2 and legacy API key auth.
servers:
  - url: https://example.de
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Legacy static bearer API key from `DREAM_DIARY_API_KEY`.
    oauth2:
      type: oauth2
      description: OAuth 2.0 Authorization Code with PKCE for Claude custom connectors.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            dreamdiary.read: Read dream entries
            dreamdiary.write: Create, update and delete dream entries
  schemas:
...
    OAuthTokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 28800
        refresh_token:
          type: string
        scope:
          type: string
          example: dreamdiary.read dreamdiary.write
    OAuthTokenError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid_grant
        error_description:
          type: string
          example: Invalid PKCE code_verifier.
...
security:
  - bearerAuth: [ ]
  - oauth2: [ dreamdiary.read, dreamdiary.write ]
paths:
...
  /oauth/token:
    post:
      operationId: issueOauthToken
      summary: OAuth token endpoint
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - grant_type
              properties:
                grant_type:
                  type: string
                  enum: [ authorization_code, refresh_token, password ]
                code:
                  type: string
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
        '400':
          description: OAuth error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenError'
        '401':
          description: OAuth client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenError'
...